# refactor: improve the smoothness of multi-select nodes when dragging