… src/plugin/selection/index.less