<!--- Provide a general summary of your changes in the Title above -->

### Description

<!--- Describe your changes in detail -->
优化前：
- 每次拖拽移动都会触发完整的"删除-重建"流程，比如对于 N 个选中的节点，每次更新需要执行 3N 次 DOM 操作（删除 + 创建 + 样式设置）
- 对于 mousemove 没有节流机制
- 拖拽时也去计算边界框，计算资源浪费

优化后：
- 更新位置时，直接更新现有 DOM 元素位置，而不是删除重建，同时只计算和更新位置差值，避免重新计算绝对位置，减少重排
- 节流：将更新频率限制在60 fps
- 使用 ```isDragging ``` 和 ```boxesUpdated``` 管理拖拽状态

| Before | After |
| ------ | ----- |
|    ![selection_before_optimizing](https://github.com/user-attachments/assets/2dce8f74-65ad-4ee6-9bf8-be7f828597bf)    |   ![selection_after_optimizing](https://github.com/user-attachments/assets/76303148-db47-43b5-b472-8286e1ebfc0b)   |

### Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
<!--- GIF or snapshot should be provided if includes UI/interactive modification. -->
<!--- How to fix the problem, and list final API implementation and usage sample if that is an new feature. -->

### Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Enhancement (changes that improvement of current feature or performance)
- [x] Refactoring (changes that neither fixes a bug nor adds a feature)
- [ ] Test Case (changes that add missing tests or correct existing tests)
- [ ] Code style optimization (changes that do not affect the meaning of the code)
- [ ] Docs (changes that only update documentation)
- [ ] Chore (changes that don't modify src or test files)

### Self Check before Merge

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the [**CONTRIBUTING**](https://github.com/antvis/x6/blob/master/CONTRIBUTING.md) document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.